+++ clang/test/Driver/pseudo-probe.c
@@ -1,16 +1,27 @@
-// RUN: %clang -### -fpseudo-probe-for-profiling %s 2>&1 | FileCheck %s --check-prefix=YESPROBE
+// RUN: %clang -### -fpseudo-probe-for-profiling %s 2>&1 | FileCheck %s --check-prefix=AUTO
 // RUN: %clang -### -fno-pseudo-probe-for-profiling %s 2>&1 | FileCheck %s --check-prefix=NOPROBE
-// RUN: %clang -### -fpseudo-probe-for-profiling -fdebug-info-for-profiling %s 2>&1 | FileCheck %s --check-prefix=YESPROBE --check-prefix=YESDEBUG  
-// RUN: %clang -### -fpseudo-probe-for-profiling -funique-internal-linkage-names %s 2>&1 | FileCheck %s --check-prefix=YESPROBE
-// RUN: %clang -### -fpseudo-probe-for-profiling -fno-unique-internal-linkage-names %s 2>&1 | FileCheck %s --check-prefix=NONAME
+// RUN: %clang -### -fpseudo-probe-for-profiling -fdebug-info-for-profiling %s 2>&1 | FileCheck %s --check-prefix=AUTO --check-prefix=YESDEBUG
+// RUN: %clang -### -fpseudo-probe-for-profiling \
+// RUN:   -funique-internal-linkage-names %s 2>&1 \
+// RUN:   | FileCheck %s --check-prefix=FUNCTIONS
+// RUN: %clang -### -fpseudo-probe-for-profiling \
+// RUN:   -funique-internal-linkage-names=all %s 2>&1 \
+// RUN:   | FileCheck %s --check-prefix=ALL
+// RUN: %clang -### -fpseudo-probe-for-profiling \
+// RUN:   -fno-unique-internal-linkage-names %s 2>&1 \
+// RUN:   | FileCheck %s --check-prefix=NONE
 
 // YESDEBUG: -fdebug-info-for-profiling
-// YESPROBE: -fpseudo-probe-for-profiling
-// YESPROBE: -funique-internal-linkage-names
+// AUTO-DAG: "-fpseudo-probe-for-profiling"
+// AUTO-DAG: "-funique-internal-linkage-names=functions"
+// FUNCTIONS-DAG: "-fpseudo-probe-for-profiling"
+// FUNCTIONS-DAG: "-funique-internal-linkage-names=functions"
+// ALL-DAG: "-fpseudo-probe-for-profiling"
+// ALL-DAG: "-funique-internal-linkage-names=all"
+// NONE-DAG: "-fpseudo-probe-for-profiling"
+// NONE-DAG: "-funique-internal-linkage-names=none"
 // NOPROBE-NOT: -fpseudo-probe-for-profiling
 // NOPROBE-NOT: -funique-internal-linkage-names
-// NONAME: -fpseudo-probe-for-profiling
-// NONAME-NOT: -funique-internal-linkage-names
 
 // On Darwin, -fpseudo-probe-for-profiling should trigger dsymutil
 // RUN: %clang -target arm64-apple-darwin -### -o foo -fpseudo-probe-for-profiling %s 2>&1 | FileCheck %s --check-prefix=CHECK-DSYMUTIL-PSEUDO-PROBE